Kane-supper-auto-forward_full_webhook_logs/requirements.txt
telethon
flask
python-dotenv
